looma_crm_future/requirements.txt
# Looma CRM AI架构重构 - 统一依赖管理
# 基于Python Sanic + 微服务架构

# Web框架
sanic==23.12.1
sanic-cors==2.2.0

# 数据库驱动
sqlalchemy==2.0.23
aiomysql==0.2.0
mysql-connector-python==9.4.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
neo4j==5.15.0
pymongo==4.15.1
motor==3.7.1

# 数据验证和序列化
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pydantic==2.5.2

# AI和机器学习
openai==1.3.7
transformers==4.36.2
torch>=2.2.0
numpy>=1.24.3
scikit-learn>=1.3.2

# 向量数据库
weaviate-client==3.25.3

# 搜索引擎
elasticsearch==8.11.0

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# 工具库
python-dotenv==1.0.0
uuid==1.30
python-dateutil==2.8.2
bcrypt==4.1.2
aiohttp==3.9.1
httpx==0.25.2

# 监控和日志
prometheus-client==0.19.0
loguru==0.7.2
psutil==5.9.6

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.5

# 测试工具
locust==2.17.0
testcontainers==3.7.1
selenium==4.15.2

# 部署工具
docker==6.1.3
docker-compose==1.29.2
